Creating the Economy controller
The economy controller class is instanciated when the terrain is loaded and is a script component of it. The EconomyController will have to do some initialisations and is able to keep track of the data that is involved in the economical part. Its exposing members that are the global settings for the simulation.

This class is only a stub now.

Signed-off-by: HaMster21 <hamster.of.dev@gmail.com>