Creating the TaskScheduler
This class is taking care of executing the tasks that each entity in the economy is producing in order. That means it will provide means to schedule tasks to run in parallel or to execute a stack of tasks to run in the order they where queued.

This class is only a stub at the moment.

Signed-off-by: HaMster21 <hamster.of.dev@gmail.com>