Adding coordinates
A producing facility is now knowing about its position on the map. This will be essential to a city that has to collect the ressources produced there with a merchant. PointF is also used for the Windows Forms to draw pictures so the producers can be easily added to a map.