Recreating the project files
The project has been recreated using Visual Studio 2010. Every unnecessary or corrupt file has been removed.

I added a town class that stores a number of producers that a created near it. A Town is created with the specified number of producers, that in turn produce ressources (weight per second). A carrier should be added to get ressources from the producer to the town. The town also needs its own stash to save ressources.